+++ activerecord/lib/active_record/relation/predicate_builder.rb
@@ -1,13 +1,14 @@
+require "active_record/relation/predicate_builder/array_handler"
+require "active_record/relation/predicate_builder/base_handler"
+require "active_record/relation/predicate_builder/basic_object_handler"
+require "active_record/relation/predicate_builder/range_handler"
+require "active_record/relation/predicate_builder/relation_handler"
+
+require "active_record/relation/predicate_builder/association_query_value"
+require "active_record/relation/predicate_builder/polymorphic_array_value"
+
 module ActiveRecord
   class PredicateBuilder # :nodoc:
-    require "active_record/relation/predicate_builder/array_handler"
-    require "active_record/relation/predicate_builder/association_query_handler"
-    require "active_record/relation/predicate_builder/base_handler"
-    require "active_record/relation/predicate_builder/basic_object_handler"
-    require "active_record/relation/predicate_builder/polymorphic_array_handler"
-    require "active_record/relation/predicate_builder/range_handler"
-    require "active_record/relation/predicate_builder/relation_handler"
-
     delegate :resolve_column_aliases, to: :table
 
     def initialize(table)
